+++ services/stadiumService.js
@@ -10,22 +10,32 @@ class StadiumsService {
   }
 
 
-  create = (req, res, next) => {
-    if (!req.body) return next(new AppError("No form data found", 404));
+  create = async (req, res, next) => {
+    var token = req.headers['authorization'];
     const values = req.body;
-    sql.query(
-      "INSERT INTO tb_estadios SET ?",
-      values,
-      function (err, data, fields) {
-        if (err) return next(new AppError(err, 500));
-        res.status(201).json({
-          status: "success",
-          message: "todo created!",
+    await autoken.adminVerificar(token).then(result => {
+      console.log(result);
+      if(result)
+      {
+        sql.query("INSERT INTO tb_estadios SET ?", values,
+          function (err, data, fields) {
+            if (err) return next(new AppError(err, 500));
+            res.status(201).json({
+              status: "success",
+              message: "Estadio creado!",
+            });
+          }
+        );
+      }else {
+        console.log(result);
+        res.status(401).send({
+        error: 'Token inválido'
         });
       }
-    );
+    }).catch(err => {
+      console.log(err);
+    })
   };
-
 
 
   findOne = async (req, res, next) => {
@@ -77,38 +87,71 @@ class StadiumsService {
   };
 
 
-  update = (req, res, next) => {
-    const body = req.body;  
-    if (!req.params.id) {
-      return next(new AppError("Id no encontrado", 404));
-    }
-    sql.query(`UPDATE tb_estadios SET nombre_estadio = ?, capacidad_estadio = ?, ciudad_estadio = ?, 
-    descripcion_estadio = ?, foto_estadio = ? WHERE id_estadio = ?`, [body.nombre_estadio, body.capacidad_estadio, body.ciudad_estadio, body.descripcion_estadio, body.foto_estadio, req.params.id],
-      function (err, data, fields) {
-        if (err) return next(new AppError(err, 500));
-        res.status(201).json({
-          status: "success",
-          message: "Estadio actualizado!",
+  update = async (req, res, next) => {
+    var token = req.headers['authorization'];
+  
+    await autoken.adminVerificar(token).then(result => {
+      console.log(result);
+      if(result) {
+        if (!req.body) {
+          return next(new AppError("No form data found", 404));
+        }
+        const body = req.body;
+        if (!req.params.id) {
+          return next(new AppError("Id no encontrado", 404));
+        }
+        sql.query(`UPDATE tb_estadios SET nombre_estadio = ?, capacidad_estadio = ?, ciudad_estadio = ?, 
+        descripcion_estadio = ?, foto_estadio = ? WHERE id_estadio = ?`, [body.nombre_estadio, body.capacidad_estadio, body.ciudad_estadio, body.descripcion_estadio, body.foto_estadio, req.params.id],
+          function (err, data, fields) {
+            if (err) return next(new AppError(err, 500));
+            res.status(201).json({
+              status: "success",
+              message: "Estadio actualizado!",
+            });
+          }
+        );
+      }else {
+        console.log(result);
+        res.status(401).send({
+        error: 'Token inválido'
         });
       }
-    );
+    }).catch(err => {
+      console.log(err);
+    })
   };
 
-      
-  delete = (req, res, next) => {
-    if (!req.params.id) {
-      return next(new AppError("No todo id found", 404));
-    }
-    sql.query(`DELETE FROM tb_estadios WHERE id_estadio = ?`, req.params.id,
-      function (err, fields) {
-        if (err) return next(new AppError(err, 500));
-        res.status(201).json({
-          status: "success",
-          message: "Estadio eliminado!",
+
+
+  delete = async (req, res, next) => {
+    var token = req.headers['authorization'];
+  
+    await autoken.adminVerificar(token).then(result => {
+      console.log(result);
+      if(result) {
+        if (!req.params.id) {
+          return next(new AppError("Estadio no encontrado!", 404));
+        }
+        sql.query(`DELETE FROM tb_estadios WHERE id_estadio = ?`, req.params.id,
+          function (err, fields) {
+            if (err) return next(new AppError(err, 500));
+            res.status(201).json({
+              status: "success",
+              message: "Estadio eliminado!",
+            });
+          }
+        );
+      }else {
+        console.log(result);
+        res.status(401).send({
+        error: 'Token inválido'
         });
       }
-    );
-  }
+    }).catch(err => {
+      console.log(err);
+    })
+  };
+
 }
 
 module.exports = StadiumsService;
